requirements.txt
anyio==4.2.0
openai==0.27.7
chromadb==0.3.26
faiss-cpu==1.7.4
tiktoken==0.4.0
langchain==0.1.4
bcrypt==4.1.2
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.86.0
fastapi-jwt-auth==0.5.0
google-api-core==2.16.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
h11==0.14.0
html2text==2020.1.16
httplib2==0.22.0
idna==3.6
jose==1.0.0
oauthlib==3.2.2
passlib==1.7.4
pdfkit==1.0.0
pillow==10.2.0
protobuf==4.25.2
psycopg2==2.9.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.14
PyJWT==1.7.1
pyparsing==3.1.1
python-dotenv==1.0.1
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.20.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.19.0
wkhtmltopdf==0.2
$2b$12$6Tjw.aDYKpD2d4H53ztobOj9tdVPd2M1OMHPrMsxm0KODj8lTUhra